Update and rename findWordPosition.java to FindWordPosition.java
